Refactor name of files
Change name to english convention
